Multi line program and pragma support
