+++ Lingo.cabal
@@ -23,7 +23,6 @@ executable Lingo
                      , Parsing.Parser
                      , Core.Data
                      , Core.Error
-                     , Core.Evaluation
                      , Core.Judgement
   default-language:    Haskell2010
   build-depends:       base >= 4.7 && < 5
